+++ gcb_web_auth/backends/oauth.py
@@ -1,4 +1,4 @@
-from ..oauth_utils import user_details_from_token, OAuthException
+from ..utils import user_details_from_token, OAuthException
 from .base import BaseBackend
 import logging
